chore: split ObjectMetadataHeaders from HeaderUtil
