+++ server/src/main/kotlin/com/adobe/testing/s3mock/s3/util/HeaderUtil.kt
@@ -20,36 +20,13 @@ import com.adobe.testing.s3mock.common.AwsHttpHeaders.AWS_CHUNKED
 import com.adobe.testing.s3mock.common.AwsHttpHeaders.X_AMZ_CHECKSUM_ALGORITHM
 import com.adobe.testing.s3mock.common.AwsHttpHeaders.X_AMZ_CHECKSUM_TYPE
 import com.adobe.testing.s3mock.common.AwsHttpHeaders.X_AMZ_CONTENT_SHA256
-import com.adobe.testing.s3mock.common.AwsHttpHeaders.X_AMZ_COPY_SOURCE_VERSION_ID
 import com.adobe.testing.s3mock.common.AwsHttpHeaders.X_AMZ_SDK_CHECKSUM_ALGORITHM
 import com.adobe.testing.s3mock.common.AwsHttpHeaders.X_AMZ_SERVER_SIDE_ENCRYPTION
-import com.adobe.testing.s3mock.common.AwsHttpHeaders.X_AMZ_STORAGE_CLASS
-import com.adobe.testing.s3mock.common.AwsHttpHeaders.X_AMZ_VERSION_ID
 import com.adobe.testing.s3mock.s3.dto.ChecksumAlgorithm
 import com.adobe.testing.s3mock.s3.dto.ChecksumType
-import com.adobe.testing.s3mock.s3.dto.CopySource
-import com.adobe.testing.s3mock.s3.dto.StorageClass
-import com.adobe.testing.s3mock.s3.model.S3ObjectMetadata
-import com.adobe.testing.s3mock.s3.util.HeaderUtil.HEADER_X_AMZ_META_PREFIX
-import com.adobe.testing.s3mock.s3.util.HeaderUtil.checksumHeaderFrom
 import org.springframework.http.HttpHeaders
 import org.springframework.http.InvalidMediaTypeException
 import org.springframework.http.MediaType
-
-fun S3ObjectMetadata.objectMetadataHeaders(
-  versioning: Boolean,
-  queryParams: Map<String, String>,
-  includeChecksum: Boolean = true,
-): Map<String, String> =
-  buildMap {
-    putAll(versionHeader(versioning))
-    storeHeaders?.let { putAll(it) }
-    putAll(userMetadataHeaders())
-    encryptionHeaders?.let { putAll(it) }
-    if (includeChecksum) putAll(checksumHeader())
-    putAll(storageClassHeaders())
-    putAll(HeaderUtil.overrideHeadersFrom(queryParams))
-  }
 
 fun resolveChecksum(
   headers: HttpHeaders,
@@ -61,50 +38,6 @@ fun resolveChecksum(
     fromSdk != null -> calculatedChecksum to fromSdk
     fromHeader != null -> HeaderUtil.checksumFrom(headers) to fromHeader
     else -> null to null
-  }
-}
-
-fun CopySource.versionHeader(versioning: Boolean): Map<String, String> =
-  if (versioning && !versionId.isNullOrEmpty()) {
-    mapOf(X_AMZ_COPY_SOURCE_VERSION_ID to versionId)
-  } else {
-    emptyMap()
-  }
-
-fun S3ObjectMetadata.versionHeader(versioning: Boolean): Map<String, String> =
-  if (versioning && !versionId.isNullOrEmpty()) {
-    mapOf(X_AMZ_VERSION_ID to versionId)
-  } else {
-    emptyMap()
-  }
-
-fun S3ObjectMetadata.checksumHeader(): Map<String, String> {
-  val checksumAlgorithm = this.checksumAlgorithm
-  val checksum = this.checksum
-  return checksumHeaderFrom(checksum, checksumAlgorithm)
-}
-
-/**
- * Creates response headers from S3ObjectMetadata StorageClass.
- */
-fun S3ObjectMetadata.storageClassHeaders(): Map<String, String> {
-  val storageClass = this.storageClass ?: return emptyMap()
-  if (storageClass == StorageClass.STANDARD) return emptyMap()
-  return mapOf(X_AMZ_STORAGE_CLASS to storageClass.toString())
-}
-
-/**
- * Creates response headers from S3ObjectMetadata user metadata.
- */
-fun S3ObjectMetadata.userMetadataHeaders(): Map<String, String> {
-  val user = this.userMetadata.orEmpty()
-  return buildMap {
-    user.forEach { (k, v) ->
-      if (k.isNotBlank() && v.isNotBlank()) {
-        val key = if (k.startsWith(HEADER_X_AMZ_META_PREFIX, ignoreCase = true)) k else HEADER_X_AMZ_META_PREFIX + k
-        put(key, v)
-      }
-    }
   }
 }
