+++ server/src/test/kotlin/com/adobe/testing/s3mock/s3/util/HeaderUtilTest.kt
@@ -16,19 +16,12 @@
 package com.adobe.testing.s3mock.s3.util
 
 import com.adobe.testing.s3mock.s3.dto.ChecksumAlgorithm
-import com.adobe.testing.s3mock.s3.dto.ChecksumType
-import com.adobe.testing.s3mock.s3.dto.Owner
-import com.adobe.testing.s3mock.s3.dto.StorageClass
-import com.adobe.testing.s3mock.s3.model.S3ObjectMetadata
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.entry
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
 import org.springframework.http.HttpHeaders
-import java.nio.file.Path
-import java.time.Instant
-import java.util.UUID
 
 internal class HeaderUtilTest {
   @ParameterizedTest
@@ -85,54 +78,6 @@ internal class HeaderUtilTest {
     val userMetadata = HeaderUtil.userMetadataFrom(httpHeaders)
     assertThat(userMetadata).containsEntry(X_AMZ_LOWERCASE_HEADER, TEST_VALUE)
   }
-
-  @Test
-  fun testCreateUserMetadata_canonical() {
-    val userMetadata = mapOf(X_AMZ_CANONICAL_HEADER to TEST_VALUE)
-    val s3ObjectMetadata = s3ObjectMetadata(userMetadata = userMetadata)
-
-    val userMetadataHeaders = s3ObjectMetadata.userMetadataHeaders()
-    assertThat(userMetadataHeaders).containsEntry(X_AMZ_CANONICAL_HEADER, TEST_VALUE)
-  }
-
-  @Test
-  fun testCreateUserMetadata_javaSdk() {
-    val userMetadata = mapOf(X_AMZ_LOWERCASE_HEADER to TEST_VALUE)
-    val s3ObjectMetadata = s3ObjectMetadata(userMetadata = userMetadata)
-
-    val userMetadataHeaders = s3ObjectMetadata.userMetadataHeaders()
-    assertThat(userMetadataHeaders).containsEntry(X_AMZ_LOWERCASE_HEADER, TEST_VALUE)
-  }
-
-  private fun s3ObjectMetadata(
-    id: UUID = UUID.randomUUID(),
-    key: String = "key",
-    userMetadata: Map<String, String>? = null,
-  ): S3ObjectMetadata =
-    S3ObjectMetadata(
-      id,
-      key,
-      "size",
-      "lastModified",
-      "\"etag\"",
-      null,
-      Instant.now().toEpochMilli(),
-      Path.of("test"),
-      userMetadata,
-      null,
-      null,
-      null,
-      Owner(0L.toString()),
-      null,
-      null,
-      ChecksumAlgorithm.SHA256,
-      null,
-      StorageClass.STANDARD,
-      null,
-      null,
-      false,
-      ChecksumType.FULL_OBJECT,
-    )
 
   companion object {
     private const val X_AMZ_CANONICAL_HEADER = "X-Amz-Meta-Some-header"
